Entire codebase rewritten to use jQuery + small fixes
I took the native JS code and make all element selection/manipulation
  use jQuery instead. This makes the code a lot cleaner in my opinion.
For separation of content, I moved all of the JS code to a new file.
I also removed most inline style tags and moved them to JS/CSS
The huge border blocks were substituted for a single border block with
  the style set to "outset", which achieves just about the same effect. If
  needed, tweak the colors to better fit the old version.
The "menutooltip" id was used multiple times, so I changed it to a class
The <hr>s were hard to follow in the menu markup so I changed them
  to <section>s and added the appropriate styles to make them look like
  the <hr> is still there.
The shop tooltip for some items was displaying "generates 0 smiles
  per second" and "You pay Infinity smiles per +1 SPS". I added a check
  to make sure that this information is only displayed when it makes sense
When an item was clicked in the shop near the edge of the screen, the
  mouse tooltip would wrap and display weirdly. I fixed this as well.
I got rid of the .padding element and applied the padding rule directly to #menu
The <div>s with the "button" class have been replaced with actual buttons
Some alignment improvements were made to the Credits modal's floated image
The redundant loading animations were removed, and the 2nd and 3rd spinner
  got the animation-delay property, which when given a negative value delay
  the animation while keeping the element visible
I also renamed "pinkie.svg" and "octavia.svg" to make the names consistent.
  Originally, I only made the files' first letters uppercase, but my Git client didn't
  recognize it as a change, so I added the ponies' last names too.